3g datang 2009 12 08 mips

+++ Data/User.cpp
@@ -3,43 +3,39 @@
 
 namespace Data
 {
-	std::string DataAccess<User>::tableName_ = "user";
-	sqlite3* DataAccess<User>::db_ = Data::GetDB();
+    std::string DataAccess<User>::tableName_ = "user";
+    sqlite3* DataAccess<User>::db_ = Data::GetDB();
     int DataAccess<User>::rowCount_ = 0;
-	int DataAccess<User>::offset_ = 0;
-	Indication DataAccess<User>::indication_ = 0;
+    int DataAccess<User>::offset_ = 0;
+    Indication DataAccess<User>::indication_ = 0;
 
-	User User::currentUser_;
+    User User::currentUser_;
 
-	User::User()
-		: DataAccess<User>()
-		//, baseInfo_(new Contact())
-		//, baseInfoId_(0)
-		//, password_(0)
-		//, setting_(new Setting())
-		//, settingId_(0)
-	{
-	}
+    User::User()
+    : DataAccess<User>()
+        //, baseInfo_(new Contact())
+        //, baseInfoId_(0)
+        //, password_(0)
+        //, setting_(new Setting())
+        //, settingId_(0)
+     {
+    }
 
-	User::~User()
-	{
-		//delete baseInfo_;
-		//delete setting_;
-	}
+    User::~User() {
+        //delete baseInfo_;
+        //delete setting_;
+    }
 
-	std::vector<boost::shared_ptr<User> > User::GetFromDatabase(std::string filter, Direction const dir, int const id, int const pageSize)
-	{
-		std::vector<boost::shared_ptr<User> > result = DataAccess<User>::GetDatasByFilter(filter, modifyFieldByDB_, dir, id, pageSize);
-		for (int i = 0; i < result.size(); ++i)
-		{
-			//result[i]->contact_ = Contact::GetFromDatabase("id = " + result[i]->contactId_);
-			//result[i]->setting_ = Setting::GetFromDatabase("id = " + result[i]->settingId_);
-		}
-		return result;
-	}
+    std::vector<boost::shared_ptr<User> > User::GetFromDatabase(std::string filter, Direction const dir, int const id, int const pageSize) {
+        std::vector<boost::shared_ptr<User> > result = DataAccess<User>::GetDatasByFilter(filter, modifyFieldByDB_, dir, id, pageSize);
+        for (size_t i = 0; i < result.size(); ++i) {
+            //result[i]->contact_ = Contact::GetFromDatabase("id = " + result[i]->contactId_);
+            //result[i]->setting_ = Setting::GetFromDatabase("id = " + result[i]->settingId_);
+        }
+        return result;
+    }
 
-	void User::Update() const
-	{
+    void User::Update() const {
         std::string cmd = "UPDATE ";
         cmd += tableName();
         cmd += " SET [baseInfoId] = ";
@@ -51,10 +47,9 @@ namespace Data
         cmd += " WHERE id = ";
         cmd += Util::StringOp::FromInt(id());
         ExecCommand(cmd);
-	}
+    }
 
-	void User::Insert()
-	{
+    void User::Insert() {
         std::string cmd = "INSERT INTO ";
         cmd += tableName();
         cmd += " ( baseInfoId, password, settingId ) VALUES ( ";
@@ -66,23 +61,20 @@ namespace Data
         cmd += " )";
         ExecCommand(cmd);
         id(GetCurrentId());
-	}
+    }
 
-    void User::Remove() const
-    {
+    void User::Remove() const {
         User::Remove("id = " + Util::StringOp::FromInt(id()));
     }
 
-	void User::Remove(std::string filter)
-	{
+    void User::Remove(std::string filter) {
         DataAccess<User>::RemoveDatasByFilter(filter);
-	}
+    }
 
-	void User::modifyFieldByDB_(int argc, char** argv, char** columnName, boost::shared_ptr<User> item)
-	{
+    void User::modifyFieldByDB_(int argc, char** argv, char** columnName, boost::shared_ptr<User> item) {
         item->id(atoi(argv[Data::getIndexByName(argc, columnName, "id")]));
         item->baseInfoId_ = atoi(argv[Data::getIndexByName(argc, columnName, "baseInfoId")]);
         item->password_ = argv[Data::getIndexByName(argc, columnName, "password")];
         item->settingId_ = atoi(argv[Data::getIndexByName(argc, columnName, "settingId")]);
-	}
+    }
 }
